requirements.txt
python-arango
python-dotenv
sentence-transformers
numpy
python-dateutil
rich
datasets
